+++ lotus/gpu/include/lotus/gpu/backends/vulkan/commands.h
@@ -70,11 +70,11 @@ namespace lotus::gpu::backends::vulkan {
 		void bind_index_buffer(const buffer&, usize offset_bytes, index_format);
 		/// Calls \p vk::CommandBuffer::bindDescriptorSets().
 		void bind_graphics_descriptor_sets(
-			const pipeline_resources&, usize first, std::span<const gpu::descriptor_set *const>
+			const pipeline_resources&, u32 first, std::span<const gpu::descriptor_set *const>
 		);
 		/// Calls \p vk::CommandBuffer::bindDescriptorSets().
 		void bind_compute_descriptor_sets(
-			const pipeline_resources&, usize first, std::span<const gpu::descriptor_set *const>
+			const pipeline_resources&, u32 first, std::span<const gpu::descriptor_set *const>
 		);
 
 		/// Calls \p vk::CommandBuffer::setViewport().
@@ -95,15 +95,10 @@ namespace lotus::gpu::backends::vulkan {
 		);
 
 		/// Calls \p vk::CommandBuffer::draw().
-		void draw_instanced(
-			usize first_vertex, usize vertex_count,
-			usize first_instance, usize instance_count
-		);
+		void draw_instanced(u32 first_vertex, u32 vertex_count, u32 first_instance, u32 instance_count);
 		/// Calls \p vk::CommandBuffer::drawIndexed().
 		void draw_indexed_instanced(
-			usize first_index, usize index_count,
-			usize first_vertex,
-			usize first_instance, usize instance_count
+			u32 first_index, u32 index_count, i32 first_vertex, u32 first_instance, u32 instance_count
 		);
 		/// Calls \p vk::CommandBuffer::dispatch().
 		void run_compute_shader(u32 x, u32 y, u32 z);
@@ -147,7 +142,7 @@ namespace lotus::gpu::backends::vulkan {
 		void bind_pipeline_state(const raytracing_pipeline_state&);
 		/// Calls \p vk::CommandBuffer::bindDescriptorSets().
 		void bind_ray_tracing_descriptor_sets(
-			const pipeline_resources&, usize first, std::span<const gpu::descriptor_set *const>
+			const pipeline_resources&, u32 first, std::span<const gpu::descriptor_set *const>
 		);
 		/// Calls \p vk::CommandBuffer::traceRaysKHR().
 		void trace_rays(
